requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
faiss-cpu
pypdf
python-dotenv
streamlit
